requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
dj-rest-auth==7.0.1
django==4.2.23
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
fakeredis==2.21.3
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.0.2
pytest-django==4.8.0
pytest-mock==3.12.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
sortedcontainers==2.4.0
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.9.0
